Some minor code refactoring.
