+++ src/main/java/org/halvors/nuclearphysics/common/science/grid/ThermalGrid.java
@@ -1,4 +1,4 @@
-package org.halvors.nuclearphysics.common.grid.thermal;
+package org.halvors.nuclearphysics.common.science.grid;
 
 import net.minecraft.block.material.Material;
 import net.minecraft.tileentity.TileEntity;
@@ -10,7 +10,6 @@ import org.halvors.nuclearphysics.api.tile.IReactor;
 import org.halvors.nuclearphysics.common.NuclearPhysics;
 import org.halvors.nuclearphysics.common.capabilities.CapabilityBoilHandler;
 import org.halvors.nuclearphysics.common.event.ThermalEvent.ThermalUpdateEvent;
-import org.halvors.nuclearphysics.common.grid.IGrid;
 import org.halvors.nuclearphysics.common.science.physics.ThermalPhysics;
 import org.halvors.nuclearphysics.common.type.Pair;
